+++ effects-cats-2/shared/src/main/scala/busymachines/pureharm/effects/PureharmEffectsSyntax.scala
@@ -14,24 +14,13 @@
  * limitations under the License.
  */
 
-package busymachines.pureharm.effects.aliases
+package busymachines.pureharm.effects
 
-import busymachines.pureharm.effects
+import busymachines.pureharm.effects.{aliases, internals}
 
-/** @author
-  *   Lorand Szakacs, https://github.com/lorandszakacs
-  * @since 13
-  *   Jun 2019
-  */
-trait PureharmEffectsTypeDefinitions {
-
-  //----------- handy custom types -----------
-  type Attempt[+R] = effects.Attempt[R]
-
-  type AttemptT[F[_], R] = effects.AttemptT[F, R]
-  val AttemptT: effects.AttemptT.type = effects.AttemptT
-
-  type Random[F[_]] = busymachines.pureharm.capabilities.Random[F]
-  val Random: busymachines.pureharm.capabilities.Random.type = busymachines.pureharm.capabilities.Random
-
-}
+// format: off
+trait PureharmEffectsSyntax
+  extends aliases.CatsSyntax
+  with internals.PureharmUnlawfulInstances
+  with internals.PureharmSyntax.Implicits
+// format: on