+++ effects-cats-2/shared/src/main/scala/busymachines/pureharm/effects/aliases/PureharmEffectsPoolAliases.scala
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package busymachines.pureharm.effects.aliases
+package busymachines.pureharm.effects
 
-import busymachines.pureharm.effects.pools
-
-trait PureharmEffectsPoolAliases {
+trait PureharmEffectPoolAliases {
+  val IORuntime: pools.IORuntime.type = pools.IORuntime
 
   val ExecutionContextFT = pools.ExecutionContextCT
 
@@ -35,6 +34,14 @@ trait PureharmEffectsPoolAliases {
   /** Denotes execution contexts backed by a cached thread pool
     */
   type ExecutionContextCT = pools.ExecutionContextCT
+
+  val ExecutionContextMainFT = pools.ExecutionContextMainFT
+
+  /** Similar to ExecutionContextFT, except that it guarantees that we have two threads, and it's specially designated
+    * as the pool on which most (most of the time all) CPU bound computation should be done in our apps, and the pool
+    * underlying instances of cats.effect.ContextShift and cats.effect.Timer
+    */
+  type ExecutionContextMainFT = pools.ExecutionContextMainFT
 
   /** Constructor object for various pools.
     */