+++ effects-cats-2/shared/src/main/scala/busymachines/pureharm/effects/aliases/PureharmEffectsTypeDefinitions.scala
@@ -31,6 +31,27 @@ trait PureharmEffectsTypeDefinitions {
   type AttemptT[F[_], R] = effects.AttemptT[F, R]
   val AttemptT: effects.AttemptT.type = effects.AttemptT
 
+  val BracketThrow: busymachines.pureharm.effects.BracketThrow.type = busymachines.pureharm.effects.BracketThrow
+
+  /** Used to block on an F[A], and ensure that all recovery and shifting back is always done.
+    *
+    * For instance, always ensure that any F[A] that talks to, say, amazon S3, is wrapped in such a
+    * {{{
+    *   blockingShifter.blockOn(S3Util.putSomething(...))
+    * }}}
+    *
+    * Libraries in the typelevel eco-system tend to already do this, so you don't need to be careful. For instance,
+    * doobie will always ensure that this is done to and from the EC that you provide specifically for accessing the DB.
+    * But you always need to double check, and be careful that you NEVER execute blocking IO on the same thread pool as
+    * the CPU bound one dedicated to your ContextShift[A]
+    */
+  type BlockingShifter[F[_]] = busymachines.pureharm.effects.BlockingShifter[F]
+
+  val BlockingShifter: busymachines.pureharm.effects.BlockingShifter.type =
+    busymachines.pureharm.effects.BlockingShifter
+
+  type PureharmIOApp = busymachines.pureharm.effects.PureharmIOApp
+
   type Random[F[_]] = busymachines.pureharm.capabilities.Random[F]
   val Random: busymachines.pureharm.capabilities.Random.type = busymachines.pureharm.capabilities.Random