+++ effects-cats-2/shared/src/main/scala/busymachines/pureharm/effects/pools/package.scala
@@ -38,4 +38,11 @@ package object pools {
   /** Denotes execution contexts backed by a cached thread pool
     */
   type ExecutionContextCT = ExecutionContextCT.Type
+
+  object ExecutionContextMainFT extends SproutSub[ExecutionContext]
+  /** Similar to ExecutionContextFT, except that it guarantees that we have two threads, and it's specially designated
+    * as the pool on which most (most of the time all) CPU bound computation should be done in our apps, and the pool
+    * underlying instances of cats.effect.ContextShift and cats.effect.Timer
+    */
+  type ExecutionContextMainFT = ExecutionContextMainFT.Type
 }