+++ effects-cats-2/shared/src/test/scala/busymachines/pureharm/effects/test/PureharmRandomTest.scala
@@ -14,24 +14,20 @@
  * limitations under the License.
  */
 
-package busymachines.pureharm.effects.aliases
+package busymachines.pureharm.effects.test
 
-import busymachines.pureharm.effects
+import busymachines.pureharm.effects._
+import munit.CatsEffectSuite
+import busymachines.pureharm.capabilities.Random
 
-/** @author
-  *   Lorand Szakacs, https://github.com/lorandszakacs
-  * @since 13
-  *   Jun 2019
-  */
-trait PureharmEffectsTypeDefinitions {
+final class PureharmRandomTest extends CatsEffectSuite {
 
-  //----------- handy custom types -----------
-  type Attempt[+R] = effects.Attempt[R]
-
-  type AttemptT[F[_], R] = effects.AttemptT[F, R]
-  val AttemptT: effects.AttemptT.type = effects.AttemptT
-
-  type Random[F[_]] = busymachines.pureharm.capabilities.Random[F]
-  val Random: busymachines.pureharm.capabilities.Random.type = busymachines.pureharm.capabilities.Random
-
+  test("generate distinct random UUIDs") {
+    val random = Random.threadLocalRandom[IO]
+    for {
+      uuid1 <- random.uuid
+      uuid2 <- random.uuid
+      _     <- IO(assert(uuid1 != uuid2))
+    } yield ()
+  }
 }